pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "log-poet"
version = "0.1.0"
description = "Transform system error messages into poetry"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Log Poet Team"}
]
keywords = ["poetry", "error-messages", "ai", "creative-writing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
log-poet = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/log-poet"
Documentation = "https://github.com/yourusername/log-poet#readme"
Repository = "https://github.com/yourusername/log-poet"
"Bug Tracker" = "https://github.com/yourusername/log-poet/issues"

[tool.setuptools]
package-dir = {"" = "."}

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.flake8]
max-line-length = 88
exclude = [".git", "__pycache__", ".venv", "venv"]